huge database update

+++ backend/cartGenerator.php
@@ -8,14 +8,14 @@
  * Time: 9:17 AM
  */
 
-include_once('../database/config.php');
-include_once('../database/database.php');
+include_once('database\Database.php');
+session_start();
+
 include_once('../frontend/menu.php');
 ?>
     <head>
         <meta charset="utf-8">
         <meta name="viewport" content="width=device-width">
-        <link rel="stylesheet" type="text/css" href="../css/cart.css">
     </head>
     <body>
         <header>
@@ -25,7 +25,7 @@ include_once('../frontend/menu.php');
         <div class="cart">
             <div class="row">Your Cart</div>
             <?php
-                if($_SESSION['loggedin'] == 1) {
+                if($_SESSION['login'] == true) {
                     $conn = DB::getConnection(DB_HOST, DB_NAME, DB_USERNAME, DB_PASSWORD);
                     $user_name = $_SESSION['username'];
                     $stmt = $conn->prepare('SELECT id_client FROM clienti WHERE email=?;');
@@ -51,57 +51,62 @@ include_once('../frontend/menu.php');
                     $stmt->execute();
                     $result_lista_cumparaturi = $stmt->get_result();
 
-                    while ($row = $result_lista_cumparaturi->fetch_assoc()) {
-                        $id_produs = $row['id_produs'];
-                        $quantity = "\"" . $row['cantitate'] . "\"";
+                    if($result_lista_cumparaturi->num_rows != 0 ){
+                        while ($row = $result_lista_cumparaturi->fetch_assoc()) {
+                            $id_produs = $row['id_produs'];
+                            $quantity = "\"" . $row['cantitate'] . "\"";
 
-                        $stmt->close();
-                        $stmt = $conn->prepare("SELECT * FROM produse WHERE id_produs=?;");
-                        $stmt->bind_param('i', $id_produs);
-                        $stmt->execute();
-                        $result = $stmt->get_result();
-                        $row = $result->fetch_assoc();
+                            $stmt->close();
+                            $stmt = $conn->prepare("SELECT * FROM produse WHERE id_produs=?;");
+                            $stmt->bind_param('i', $id_produs);
+                            $stmt->execute();
+                            $result = $stmt->get_result();
+                            $row = $result->fetch_assoc();
 
-                        $name = $row['nume'];
-                        $price = $row['pret'];
-                        $arome = $row['arome'];
-                        $img_path = "\"" . "../backend/products/" . $row['path_poza'] . "\"";
-                        $sour = $row['acidulat'] == 1 ? 'acidulat' : 'neacidulat';
-                        ?>
-                        <div class="column">
-                            <div class="item">
-                                <div class="buttons">
-                                    <a class="delete-btn" href="delete_from_cart"></a>
-                                    <span class="like-btn"></span>
+                            $name = $row['nume'];
+                            $price = $row['pret'];
+                            $arome = $row['arome'];
+                            $img_path = "\"" . "../backend/products/" . $row['path_poza'] . "\"";
+                            $sour = $row['acidulat'] == 1 ? 'acidulat' : 'neacidulat';
+                            ?>
+                            <div class="column">
+                                <div class="item">
+                                    <div class="buttons">
+                                        <a class="delete-btn" href="delete_from_cart"></a>
+                                        <span class="like-btn"></span>
+                                    </div>
+
+                                    <div class="image">
+                                        <img src=<?php echo $img_path ?> alt=""/>
+                                    </div>
+
+                                    <div class="description">
+                                        <span><?php echo $name ?></span>
+                                        <span><?php echo $arome ?></span>
+                                        <span><?php echo $sour ?></span>
+                                    </div>
+
+                                    <div>
+                                        <input type="number" name="quantity" value=<?php echo $quantity ?>>
+                                    </div>
+
+                                    <div class="total-price"><?php echo $price ?> Lei</div>
                                 </div>
-
-                                <div class="image">
-                                    <img src=<?php echo $img_path ?> alt=""/>
-                                </div>
-
-                                <div class="description">
-                                    <span><?php echo $name ?></span>
-                                    <span><?php echo $arome ?></span>
-                                    <span><?php echo $sour ?></span>
-                                </div>
-
-                                <div>
-                                    <input type="number" name="quantity" value=<?php echo $quantity ?>>
-                                </div>
-
-                                <div class="total-price"><?php echo $price ?> Lei</div>
                             </div>
+
+                        <div class="checkout-btn">
+                            <span><a href="../frontend/checkout.php">Check out</a></span>
                         </div>
-                        <?php
+                            <?php
+                        }
+                    } else{
+                        echo "YOUR CART IS EMPTY!";
                     }
                 } else{
                     header("Location: ../frontend/index.php");
                     die();
                 }
             ?>
-        <div class="checkout-btn">
-            <span><a href="../frontend/checkout.php">Check out</a></span>
-        </div>
         </div>
     </body>
 </html>
